weather_agent/requirements.txt
google-cloud-aiplatform[adk,agent_engines]
requests
python-dotenv
httpx>=0.26.0
# chainlit
streamlit